requirements.txt
Django==4.1.7
channels==4.0.0
react==4.3.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
pylint-django==2.5.3